refactor(data-theme): lift compose/theme payload schema to data-theme-core
Extract ThemePayload, ResolvedThemeTokens, ThemeConsumer, TypographyToken,
KIND, and SCHEMA_VERSION out of ThemeDataProduct.kt into a new pure-data
data-theme-core module so MCP clients and other connectors can depend on
the schema without pulling in Compose Material3, the daemon-side
ThemeDataProductRegistry, or the slot-table inspection facade.

ThemeDataProductRegistry.KIND / .SCHEMA_VERSION stay as forwarding
constants pointing at Material3ThemeProduct so existing in-tree call
sites (ThemeCaptureExtension, MaterialThemeTokenCaptureTest) keep
working without churn. RENDER_MODE stays where it is — it's a runtime
trigger, not part of the wire schema.

Cross-module consumers (daemon/desktop, daemon/android,
ThemeDataProductRegistryTest) gain explicit imports for the moved
types since they previously resolved them via same-package
ee.schimke.composeai.daemon visibility.

Mirrors the core/connector split that fonts, strings, layoutinspector,
resources, a11y, and (after #804) history already use. No behavior
change.
